feat(DirectMode): Add direct mode + clean test code (#20)
* feat(DirectMode): Add direct mode in s3-agent

* feat(DirectMode): Refacto tests utils + add test for rebuild mode

* refactor(tests): Clean test code

* refactor(tests): remove useless line

+++ s3-agent/tests/conftest.py
@@ -23,5 +23,6 @@ def handle_agent(request):
     yield connection.cursor()
 
     ### TEARDOWN ###
-    stop_agent(process, connection)
-    run_command(f'rm -rf {S3_AGENT_PATH}', code=0)
+    stop_agent(process, connection, not DEBUG)
+    if not DEBUG:
+        run_command(f'rm -rf {S3_AGENT_PATH}', code=0)
